requirements.txt
flask==3.0.0
boto3==1.34.34
python-dotenv==1.0.0
Pillow==10.2.0
PyPDF2==3.0.1
werkzeug==3.0.1
requests==2.31.0